Organize store a bit
