+++ src/app/components/StateEditor.js
@@ -3,9 +3,9 @@ import { connect } from 'react-redux'
 
 let StateEditor = ({ dispatch }) => {
     return (
-        <div>
-            <button onClick={() => dispatch({type: 'TOGGLE_MOVE_Y'})}>Toggle Y Movement</button>
-        </div>
+      <div>
+        <button onClick={() => dispatch({type: 'TOGGLE_MOVE_Y'})}>Toggle Y Movement</button>
+      </div>
     )
 }
