+++ src/app/store.js
@@ -1,43 +1,7 @@
 import { createStore } from 'redux';
+import dungeonReducer from './reducers/dungeonReducer.js'
 
-const store = createStore((state = {}, action) => {
-    switch (action.type) {
-      case 'TOGGLE_MOVE_Y':
-        return {
-          ...state,
-          moveY: !state.moveY
-        };
-      case 'MOUSE_DOWN':
-        return {
-          ...state,
-          mouseDown: true,
-          mouseStartX: action.x,
-          mouseStartY: action.y
-        };
-      case 'MOUSE_UP':
-        return {
-          ...state,
-          mouseDown: false
-        };
-      case 'ADD_SPACE':
-        let spaceArray = state.dungeon.spaces.slice();
-        spaceArray = [...spaceArray, action.newSpace ];
-        return {
-          ...state,
-          dungeon: {
-            ...state.dungeon,
-            spaces: spaceArray
-          }
-        };
-      case 'SELECT_TOOL':
-        return {
-          ...state,
-          selectedTool: action.selectedTool
-        };
-      default:
-        return state
-    }
-  },
+const store = createStore(dungeonReducer,
   {
     id: '823cc811-9499-4f3d-abeb-941d2ee4fd98',
     moveY: false,
